+++ packages/date-select/index.js
@@ -1,7 +1,0 @@
-import DateSelect from './src/main';
-
-DateSelect.install = function(Vue) {
-    Vue.component(DateSelect.name, DateSelect);
-};
-
-export default DateSelect;