+++ packages/form-steps/index.js
@@ -1,7 +1,0 @@
-import FormSteps from './src/main';
-
-FormSteps.install = function(Vue) {
-  Vue.component(FormSteps.name, FormSteps);
-};
-
-export default FormSteps;