+++ packages/steps/index.js
@@ -1,0 +1,7 @@
+import Steps from './src/main';
+
+Steps.install = function(Vue) {
+    Vue.component(Steps.name, Steps);
+};
+
+export default Steps;