+++ packages/tabs/src/main.vue
@@ -137,25 +137,12 @@
 <script>
 import create from '../../utils/create';
 import { formInitVal } from '../../utils/util';
+import fun from '../../mixins/fun'
 import { setTimeout } from 'timers';
 export default create({
   name: 'tabs',
+  mixins: [fun()],
   props: {
-    tableLoading: {
-      type: Boolean,
-      default: false,
-    },
-    beforeOpen: Function,
-    beforeClose: Function,
-    uploadBefore: Function,
-    uploadAfter: Function,
-    rowClassName: Function,
-    spanMethod: Function,
-    summaryMethod: Function,
-    data: {
-      type: Array,
-      default: () => []
-    },
     value: {
       type: Object,
       default: () => { }
@@ -239,54 +226,6 @@ export default create({
     submit () {
       this.$emit('submit', this.tableForm);
     },
-    refreshChange (params) {
-      this.$emit('refresh-change', params);
-    },
-    searchReset () {
-      this.$emit('search-reset');
-    },
-    sizeChange (val) {
-      this.$emit('size-change', val);
-    },
-    dateChange (val) {
-      this.$emit('date-change', val);
-    },
-    currentChange (val) {
-      this.$emit('current-change', val);
-    },
-    currentRowChange (currentRow, oldCurrentRow) {
-      this.$emit('current-row-change', currentRow, oldCurrentRow);
-    },
-    selectionChange (val) {
-      this.$emit('selection-change', val);
-    },
-    sortChange (val) {
-      this.$emit('sort-change', val);
-    },
-    searchChange (val) {
-      this.$emit('search-change', val);
-    },
-    rowDblclick (row, event) {
-      this.$emit('row-dblclick', row, event);
-    },
-    rowClick (row, event, column) {
-      this.$emit('row-click', row, event, column);
-    },
-    rowDel (row, index) {
-      this.$emit('row-del', row, index);
-    },
-    rowSave (row, done, loading) {
-      this.$emit('row-save', row, done, loading);
-    },
-    rowUpdate (row, index, done, loading) {
-      this.$emit('row-update', row, index, done, loading);
-    },
-    rowAdd () {
-      this.$refs.crud.rowAdd();
-    },
-    rowEdit (row, index) {
-      this.$refs.crud.rowEdit(row, index);
-    }
   }
 });
 </script>